Separate clients per domain
