+++ ui/src/clients/usersClient.js
@@ -19,16 +19,6 @@ export async function createUser(email, phone) {
     return response.data;
 }
 
-export async function getUserJobs(userId) {
-    let response = await fetch(`/api/users/${userId}/jobs`);
-
-    if (response.status === 200) {
-        let data = await response.text();
-        return data
-    }
-    return response.data;
-}
-
 export async function getUserAccounts(userId) {
     
     let response = await fetch(`/api/users/${userId.userId}/accounts`);
@@ -54,24 +44,6 @@ export async function getUserAccount(url) {
         body: body
     });
 
-    if (response.status === 200) {
-        let data = await response.text();
-        return data
-    }
-    
-    return response.data;
-}
-
-
-export async function getJob(jobId) {
-    let response = await fetch(`https://au-api.basiq.io/jobs/${jobId}`, {
-        method: 'GET',
-        headers: {
-            'Content-Type': 'application/json',
-            'Authorization': `Bearer ${sessionStorage.getItem("session_token")}`
-        }
-    });
-    
     if (response.status === 200) {
         let data = await response.text();
         return data
